# src/resources/npm/api-client/src/models/ErrorUnauthenticated.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BMLT - OpenAPI 3.0
 * BMLT Admin API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ErrorUnauthenticated
 */
export interface ErrorUnauthenticated {
    /**
     * 
     * @type {string}
     * @memberof ErrorUnauthenticated
     */
    message?: string;
}

/**
 * Check if a given object implements the ErrorUnauthenticated interface.
 */
export function instanceOfErrorUnauthenticated(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function ErrorUnauthenticatedFromJSON(json: any): ErrorUnauthenticated {
    return ErrorUnauthenticatedFromJSONTyped(json, false);
}

export function ErrorUnauthenticatedFromJSONTyped(json: any, ignoreDiscriminator: boolean): ErrorUnauthenticated {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'message': !exists(json, 'message') ? undefined : json['message'],
    };
}

export function ErrorUnauthenticatedToJSON(value?: ErrorUnauthenticated | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'message': value.message,
    };
}