# src/resources/npm/api-client/src/models/FormatTranslation.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BMLT - OpenAPI 3.0
 * BMLT Admin API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FormatTranslation
 */
export interface FormatTranslation {
    /**
     * 
     * @type {string}
     * @memberof FormatTranslation
     */
    key: string;
    /**
     * 
     * @type {string}
     * @memberof FormatTranslation
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof FormatTranslation
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof FormatTranslation
     */
    language: string;
}

/**
 * Check if a given object implements the FormatTranslation interface.
 */
export function instanceOfFormatTranslation(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "key" in value;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "description" in value;
    isInstance = isInstance && "language" in value;

    return isInstance;
}

export function FormatTranslationFromJSON(json: any): FormatTranslation {
    return FormatTranslationFromJSONTyped(json, false);
}

export function FormatTranslationFromJSONTyped(json: any, ignoreDiscriminator: boolean): FormatTranslation {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'key': json['key'],
        'name': json['name'],
        'description': json['description'],
        'language': json['language'],
    };
}

export function FormatTranslationToJSON(value?: FormatTranslation | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'key': value.key,
        'name': value.name,
        'description': value.description,
        'language': value.language,
    };
}