# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BMLT - OpenAPI 3.0",
        "description": "BMLT Admin API Documentation",
        "license": {
            "name": "MIT",
            "url": "https://github.com/bmlt-enabled/bmlt-root-server/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/formats": {
            "get": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Retrieves formats",
                "description": "Retrieve formats",
                "operationId": "getFormats",
                "responses": {
                    "200": {
                        "description": "Returns when user is authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FormatCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Creates a format",
                "description": "Creates a format.",
                "operationId": "createFormat",
                "requestBody": {
                    "description": "Pass in format object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FormatCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Returns when POST is successful.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Format"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorIncorrectCredentials"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when user is unauthorized to perform action.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no format exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoFormatExists"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            }
        },
        "/api/v1/formats/{formatId}": {
            "get": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Retrieves a format",
                "description": "Retrieve a format",
                "operationId": "getFormat",
                "parameters": [
                    {
                        "name": "formatId",
                        "in": "path",
                        "description": "ID of format",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns when user is authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Format"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no format exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoFormatExists"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Updates a format",
                "description": "Updates a format.",
                "operationId": "updateFormat",
                "parameters": [
                    {
                        "name": "formatId",
                        "in": "path",
                        "description": "ID of format",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Pass in format object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FormatUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Success."
                    },
                    "401": {
                        "description": "Returns when user is not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorIncorrectCredentials"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when user is unauthorized to perform action.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no format exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoFormatExists"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Deletes a format",
                "description": "Deletes a format by id.",
                "operationId": "deleteFormat",
                "parameters": [
                    {
                        "name": "formatId",
                        "in": "path",
                        "description": "ID of format",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success."
                    },
                    "401": {
                        "description": "Returns when not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no format exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoFormatExists"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Patches a format",
                "description": "Patches a single format by id.",
                "operationId": "patchFormat",
                "parameters": [
                    {
                        "name": "formatId",
                        "in": "path",
                        "description": "ID of format",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Pass in fields you want to update.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FormatPartialUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Success."
                    },
                    "401": {
                        "description": "Returns when not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no format exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoFormatExists"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorIncorrectCredentials": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The provided credentials are incorrect."
                    }
                },
                "type": "object"
            },
            "ErrorUnauthenticated": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "ErrorUnauthorized": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The field is required. (and 1 more error)"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "error details"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "FormatBase": {
                "properties": {
                    "worldId": {
                        "type": "string",
                        "example": "string"
                    },
                    "type": {
                        "type": "string",
                        "example": "string"
                    },
                    "translations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FormatTranslation"
                        }
                    }
                },
                "type": "object"
            },
            "FormatTranslation": {
                "required": [
                    "key",
                    "name",
                    "description",
                    "language"
                ],
                "properties": {
                    "key": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "language": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Format": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/FormatBase"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": "0"
                            }
                        }
                    }
                ]
            },
            "FormatCreate": {
                "required": [
                    "translations"
                ],
                "allOf": [
                    {
                        "$ref": "#/components/schemas/FormatBase"
                    }
                ]
            },
            "FormatUpdate": {
                "required": [
                    "translations"
                ],
                "allOf": [
                    {
                        "$ref": "#/components/schemas/FormatBase"
                    }
                ]
            },
            "FormatPartialUpdate": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/FormatBase"
                    }
                ]
            },
            "FormatCollection": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/Format"
                }
            },
            "NoFormatExists": {
                "description": "Returns when no format exists.",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bmltToken": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "api/v1/auth/token",
                        "refreshUrl": "api/v1/auth/refresh",
                        "scopes": {}
                    }
                }
            }
        }
    }
}